Split out url generation functionality
This allows me to keep the main `InternetArchive` class focused and test a few areas I wasn't able to test before.
