+++ InternetArchiveKit/InternetArchiveProtocols.swift
@@ -21,9 +21,28 @@ public protocol InternetArchiveProtocol {
               completion: @escaping (InternetArchive.SearchResponse?, Error?) -> Void)
   func itemDetail(identifier: String,
                   completion: @escaping (InternetArchive.Item?, Error?) -> Void)
+  @available(*, deprecated, message: "Use InternetArchive.URLGenerator instead")
+  func generateItemImageUrl(itemIdentifier: String) -> URL?
+  @available(*, deprecated, message: "Use InternetArchive.URLGenerator instead")
+  func generateMetadataUrl(identifier: String) -> URL?
+  @available(*, deprecated, message: "Use InternetArchive.URLGenerator instead")
+  func generateDownloadUrl(itemIdentifier: String, fileName: String) -> URL?
+}
+
+/**
+ A protocol to which the main `InternetArchive.URLGenerator` class conforms
+ */
+public protocol InternetArchiveURLGeneratorProtocol {
   func generateItemImageUrl(itemIdentifier: String) -> URL?
   func generateMetadataUrl(identifier: String) -> URL?
   func generateDownloadUrl(itemIdentifier: String, fileName: String) -> URL?
+  // swiftlint:disable:next function_parameter_count
+  func generateSearchUrl(query: InternetArchiveURLStringProtocol,
+                         page: Int,
+                         rows: Int,
+                         fields: [String],
+                         sortFields: [InternetArchiveURLQueryItemProtocol],
+                         additionalQueryParams: [URLQueryItem]) -> URL?
 }
 
 /**
